Rewrote AsyncStatusIndicator in ts, removed the d.ts and index.js
