+++ src/AsyncStatusIndicator/AsyncStatusIndicator.tsx
@@ -1,16 +1,33 @@
-import React from 'react';
-import { skinnable, props, t } from '../utils';
+import * as React from 'react';
+import { Type } from 'tcomb';
+import { props, t } from '../utils';
 import cx from '../utils/classnames';
 import View from 'react-flexview';
 
-export const AsyncStatusIndicatorState = t.enums.of([
+export namespace AsyncStatusIndicatorProps {
+  export type AsyncStatusIndicatorState = 'ready' | 'processing' | 'success' | 'error';
+}
+
+export type AsyncStatusIndicatorProps = {
+  state: AsyncStatusIndicatorProps.AsyncStatusIndicatorState,
+  className?: string,
+  style?: React.CSSProperties,
+  icons: { [key in AsyncStatusIndicatorProps.AsyncStatusIndicatorState]?: any },
+  labels: { [key in AsyncStatusIndicatorProps.AsyncStatusIndicatorState]?: string },
+}
+
+export const AsyncStatusIndicatorState: AsyncStatusIndicatorProps.AsyncStatusIndicatorState = t.enums.of([
   'ready',
   'processing',
   'success',
   'error'
 ], 'AsyncStatusIndicatorState');
 
-export const Props = {
+export type TProps = {
+  [key: string]: Type<any> | AsyncStatusIndicatorProps.AsyncStatusIndicatorState
+};
+
+export const Props: TProps = {
   state: AsyncStatusIndicatorState,
   icons: t.dict(AsyncStatusIndicatorState, t.ReactElement),
   labels: t.dict(AsyncStatusIndicatorState, t.String),
@@ -25,18 +42,13 @@ export const Props = {
  * @param className - an optional class name to pass to top level element of the component
  * @param style - an optional style object to pass to top level element of the component
  */
-@skinnable()
 @props(Props)
-export default class AsyncStatusIndicator extends React.PureComponent {
-
-  getLocals({ state, icons, labels, className, style }) {
+export default class AsyncStatusIndicator extends React.PureComponent<AsyncStatusIndicatorProps> {
+  render() {
+    const { state, icons, labels, className, style } = this.props;
     const icon = icons[state] || null;
     const label = labels[state] || '';
 
-    return { state, icon, label, className, style };
-  }
-
-  template({ state, icon, label, className, style }) {
     return (
       <View vAlignContent='center' className={cx('async-status-indicator', state, className)} style={style}>
