--- src/AsyncStatusIndicator/index.js
@@ -1,4 +1,0 @@
-import AsyncStatusIndicator, { Props as AsyncStatusIndicatorProps, AsyncStatusIndicatorState } from './AsyncStatusIndicator';
-
-export { AsyncStatusIndicatorProps, AsyncStatusIndicatorState };
-export default AsyncStatusIndicator;
